+++ CHANGELOG.md
@@ -33,6 +33,12 @@ To install or update CocoaPods see this [guide](http://docs.cocoapods.org/guides
   [Samuel Ford](https://github.com/samuelwford)
   [#1042](https://github.com/CocoaPods/CocoaPods/issues/1042)
 
+##### Bug Fixes
+
+* Support HTTP redirects when linting homepage and screenshots.
+  [Boris Bügling](https://github.com/neonichu)
+  [#2027](https://github.com/CocoaPods/CocoaPods/pull/2027)
+
 ## 0.31.1
 [CocoaPods](https://github.com/CocoaPods/CocoaPods/compare/0.31.1...0.31.0)
 • [CocoaPods-Core](https://github.com/CocoaPods/Core/compare/0.31.1...0.31.0)